awscdk-packages/examples/requirements.txt
--extra-index-url https://${FURY_AUTH}:@pypi.fury.io/<your-company>/
infra-awscdk-devops-packages==0.1.53
aws-cdk-lib==2.160.0
constructs>=10.0.0,<11.0.0
